requirements.txt
openai
pydantic
numpy
scikit-learn
ddgs
streamlit
python-dotenv